app/requirements.txt
keras==2.3.0
nltk==3.4.5
Pillow==6.2.2
tensorflow==1.15.2
protobuf==3.20.1
